mind_tracker_bot_replit/requirements.txt

python-telegram-bot==20.3
pandas
matplotlib
